Move docking infrastructure to separate namespace
