+++ Bonsai.Editor/Docking/WorkflowDockContent.resx
