+++ Bonsai.Editor/Docking/WorkflowDockPaneStripFactory.cs
@@ -1,6 +1,6 @@
 ﻿using WeifenLuo.WinFormsUI.Docking;
 
-namespace Bonsai.Editor.GraphView
+namespace Bonsai.Editor.Docking
 {
     internal class WorkflowDockPaneStripFactory : DockPanelExtender.IDockPaneStripFactory
     {
