+++ Bonsai.Editor/Docking/WorkflowFloatWindow.cs
@@ -4,7 +4,7 @@ using System.Windows.Forms;
 using Bonsai.Design;
 using WeifenLuo.WinFormsUI.Docking;
 
-namespace Bonsai.Editor.GraphView
+namespace Bonsai.Editor.Docking
 {
     internal class WorkflowFloatWindow : FloatWindow
     {
